decode dataset json message addsample code for diferent field encoding:Raw, Variant and DataValue JsonEncode/Decoder  classes were changed to support the decoding and the correct encoding

+++ Applications/ConsoleReferencePublisher/PublishedValuesWrites.cs
@@ -235,8 +235,8 @@ namespace Quickstarts.ConsoleReferencePublisher
                 case BuiltInType.Byte:
                     if (variable.ValueRank == ValueRanks.Scalar)
                     {
-                        Byte byteValue = Convert.ToByte(dataValue.Value);
-                        dataValue.Value = byteValue + 1;
+                        byte byteValue = Convert.ToByte(dataValue.Value);
+                        dataValue.Value = (byte)(byteValue + 1);
                         valueUpdated = true;
                     }
                     break;
